Need to be able to move tmpfiles. Also we can now specify unit test suite prefix strings.

+++ src/system/tmpfile.cc
@@ -32,7 +32,13 @@ namespace hpc {
 
    tmpfile::~tmpfile()
    {
-      if( fs::exists( _path ) )
+      close();
+   }
+
+   void
+   tmpfile::close()
+   {
+      if( !_path.empty() && fs::exists( _path ) )
          fs::remove( _path );
    }
